% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collateral.R
\docType{package}
\name{collateral}
\alias{collateral}
\alias{collateral-package}
\title{collateral: Reproducible, comfortable and fast processing with R}
\description{
\code{collateral} provides you with functions you might miss in R.
}
\details{
\strong{Overview:}
\itemize{
  \item \code{p*apply()} functions for reproducible parallel processing.
  \item \code{\%onerror\%} operators for handling errors more direct.
}

\strong{However, this package is a proof of concept and please use it,
at the moment, for testing purposes only}
}
