% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r
\name{collateral_extras}
\alias{collateral_extras}
\alias{safely_mapped}
\alias{quietly_mapped}
\alias{as_safely_mapped}
\alias{as_quietly_mapped}
\alias{c.safely_mapped}
\alias{c.quietly_mapped}
\alias{[.safely_mapped}
\alias{[.quietly_mapped}
\alias{format.safely_mapped}
\alias{format.quietly_mapped}
\alias{print.safely_mapped}
\alias{print.quietly_mapped}
\title{Additional functions required to integrate
\code{\link{map_safely}} and \code{\link{map_quietly}} with
the \code{pillar} package. See the
\url{https://cran.r-project.org/web/packages/tibble/vignettes/extending.html}{extending tibbles vignette}.}
\usage{
safely_mapped(x)

quietly_mapped(x)

as_safely_mapped(x)

as_quietly_mapped(x)

\method{c}{safely_mapped}(x, ...)

\method{c}{quietly_mapped}(x, ...)

\method{[}{safely_mapped}(x, i)

\method{[}{quietly_mapped}(x, i)

\method{format}{safely_mapped}(x, ...)

\method{format}{quietly_mapped}(x, ...)

\method{print}{safely_mapped}(x, ...)

\method{print}{quietly_mapped}(x, ...)
}
\description{
Additional functions required to integrate
\code{\link{map_safely}} and \code{\link{map_quietly}} with
the \code{pillar} package. See the
\url{https://cran.r-project.org/web/packages/tibble/vignettes/extending.html}{extending tibbles vignette}.
}
\keyword{internal}
