% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.r
\name{collateral_mappers}
\alias{collateral_mappers}
\alias{map_safely}
\alias{map_quietly}
\title{Map safely or quietly over a list.}
\usage{
map_safely(.x, .f, ...)

map_quietly(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula or atomic vector.}

\item{...}{Other arguments supplied to \code{\link[purrr:map]{map}}}
}
\value{
A list of the same length as \code{.x}. The list elemnts contain
  results and captured side effects as described in
  \code{\link[purrr:safely]{safely}} and
  \code{\link[purrr:quietly]{quietly}}.
}
\description{
\code{map_safely} and \code{map_quietly} are variants of
\code{\link[purrr:map]{map}} that:
\enumerate{
  \item wrap the supplied function \code{.f} with either
    \code{\link[purrr:safely]{safely}} or
    \code{\link[purrr:quietly]{quietly}}, and
  \item add a class to the returned output list in order to format it
    nicely when it (or a tibble it appears in) is printed.
 }
}
\details{
\code{map_safely} will summarise the returned list with a fixed-width
 string of two (spaced) columns:
 \enumerate{
   \item If a \code{result} component is present, \code{R} appears, and
   \item If an \code{error} component is present, \code{E} appears.
 }
 If either component is missing, an underscore (\code{_}) appears in its
 place.

Similarly, \code{map_quietly} will summarise the returned list with a
 fixed-width string of four (spaced) columns:
 \enumerate{
   \item If a \code{result} component is present, \code{R} appears,
   \item If an \code{output} component is present, \code{O} appears,
   \item If a \code{messages} component is present, \code{M} appears, and
   \item If a \code{warnings} component is present, \code{W} appears.
 }
 If any is missing, an underscore (\code{_}) appears in its
 place.
}
