% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_here.R
\name{from_here}
\alias{from_here}
\title{Set up parallel environment from here}
\usage{
from_here(progress = FALSE, title = "Progress", memo = FALSE,
  resume = FALSE, eta = FALSE, time = FALSE, threads = 1,
  sameSeed = FALSE, stopOnError = TRUE)
}
\description{
Set up parallel environment from here.
}
\details{
Set global options.
}
\seealso{
\code{\link{plapply}} and \code{\link{psapply}}.
}
