% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onerror.R
\name{\%onerror\%}
\alias{\%onerror\%}
\title{Try blocks for errors}
\usage{
expr \%onerror\% err
}
\description{
Try expression blocks
             Use 'stop(e)' to throw error and stop.
             Use 'message(e)' to just throw the error message.
}
