% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psapply.R
\name{psapply}
\alias{psapply}
\title{Parallel sapply wrapper function}
\usage{
psapply(X, FUN, ...)
}
\description{
Parallel sapply wrapper function for plapply(..., simplify = TRUE).
             \strong{Proof of concept - for testing purposes only}
}
