% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pillar.r
\name{type_sum.safely_mapped}
\alias{type_sum.safely_mapped}
\title{#' @importFrom pillar is_vector_s3
#' @export
is_vector_s3.collat_safe <- function(x) TRUE}
\usage{
type_sum.safely_mapped(x)
}
\description{
#' @importFrom pillar is_vector_s3
#' @export
is_vector_s3.collat_quiet <- function(x) TRUE
#' @importFrom pillar obj_sum
#' @export
obj_sum.collat_safe <- function(x) { "safely" }
}
\details{
#' @importFrom pillar obj_sum
#' @export
obj_sum.collat_quiet <- function(x) { "quietly" }
}
