% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collateral.R
\docType{package}
\name{collateral}
\alias{collateral}
\alias{collateral-package}
\title{collateral: Reproducible, comfortable and fast processing with R}
\description{
\code{collateral} provides you with functions you might miss in R.
}
\details{
\strong{Overview:}
\itemize{
  \item \code{p*apply()} functions for reproducible parallel processing.
  \item \code{\%onerror\%}, \code{\%onwarning\%} operators for handling errors in a more common way.
  \item \code{throw} let you \emph{throw} errors which can contain arbitrary objects - similar to Java.
}

\strong{However, this package is a proof of concept and please use it,
at the moment, for testing purposes only}
}
\seealso{
\code{\link{plapply}}, \code{\link{\%onerror\%}} and \code{\link{throw}}
}
